Moved "Minimap.hpp" to the right folder
and small optimisation on "Minimap.cpp"
